requirements.txt
kafka-python
redis
numpy